requirements.txt
aiogram==3.18.0
pydantic_settings==2.7.1
pydantic==2.10.6
python-dotenv==1.0.1
validators==0.34.0
motor==3.7.0
pymongo==4.11.1
aiohttp_socks==0.10.1